requirements.txt
Flask==2.3.4
Flask-Login==0.7.0
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.24
python-dotenv==1.0.0
Werkzeug==3.0.3
gunicorn==21.2.0
bcrypt==4.0.1
Pillow==10.0.0
boto3==1.30.0
